refactor satellite methods into their own file
