+++ setup.cfg
@@ -22,7 +22,6 @@ disable=abstract-method, # raises false positives in beam
         no-member, # raises false positives
         no-name-in-module, # raises false positives
         too-many-arguments, # better handled by reviewer judgement
-        too-many-lines, # TODO remove
         too-many-locals, # better handled by reviewer judgement
         too-few-public-methods, # bad rule
         too-many-public-methods, # bad rule
